Add backend indicator function filter active first
added a function for the backend indicator to show actived backend
first.

changed option format for `company-posframe-backend-format-function`
new a choice wich defaults to the new function.

changed `company-posframe-quickhelp-show` to call
`company-posframe-backend-format-function` without mapping backends
the format function now takes care of that. Allowing it to order
backends in any order.