Fix x-pos if company margin extends to left of beginning of line
Do not limit the position to `(line-beginning-position)', but use point
always in same line, and then shift left according to `:company-margin'.

This keeps the posframe correctly aligned even if it extends to the left
of the window (if there's room left), and even more importantly, when
there is a `line-prefix' (as in `org-indent-mode') or a left margin in
the parent window.